Add convergence test for Poisson
