+++ Convergence/Poisson/convergence2d.inputs
@@ -75,18 +75,18 @@ FieldSolverMultigrid.plt_vars          = phi rho E Esol    # Plot variables: 'ph
 FieldSolverMultigrid.use_regrid_slopes = true              # Use slopes when regridding or not
 FieldSolverMultigrid.kappa_source      = true              # Volume weighted space charge density or not (depends on algorithm)
 
-FieldSolverMultigrid.gmg_verbosity     = -1                # GMG verbosity
+FieldSolverMultigrid.gmg_verbosity     = 10                # GMG verbosity
 FieldSolverMultigrid.gmg_pre_smooth    = 12                # Number of relaxations in downsweep
 FieldSolverMultigrid.gmg_post_smooth   = 12                # Number of relaxations in upsweep
 FieldSolverMultigrid.gmg_bott_smooth   = 12                # Number of at bottom level (before dropping to bottom solver)
 FieldSolverMultigrid.gmg_min_iter      = 5                 # Minimum number of iterations
 FieldSolverMultigrid.gmg_max_iter      = 32                # Maximum number of iterations
-FieldSolverMultigrid.gmg_exit_tol      = 1.E-10            # Residue tolerance
+FieldSolverMultigrid.gmg_exit_tol      = 1.E-14            # Residue tolerance
 FieldSolverMultigrid.gmg_exit_hang     = 0.2               # Solver hang
-FieldSolverMultigrid.gmg_min_cells     = 16                # Bottom drop
-FieldSolverMultigrid.gmg_bc_order      = 2                 # Boundary condition order for multigrid
+FieldSolverMultigrid.gmg_min_cells     = 8                 # Bottom drop
+FieldSolverMultigrid.gmg_bc_order      = 1                 # Boundary condition order for multigrid
 FieldSolverMultigrid.gmg_bc_weight     = 2                 # Boundary condition weights (for least squares)
-FieldSolverMultigrid.gmg_jump_order    = 2                 # Boundary condition order for jump conditions
+FieldSolverMultigrid.gmg_jump_order    = 1                 # Boundary condition order for jump conditions
 FieldSolverMultigrid.gmg_jump_weight   = 2                 # Boundary condition weight for jump conditions (for least squares)
 FieldSolverMultigrid.gmg_bottom_solver = bicgstab          # Bottom solver type. 'simple', 'bicgstab', or 'gmres'
 FieldSolverMultigrid.gmg_cycle         = vcycle            # Cycle type. Only 'vcycle' supported for now. 
@@ -121,8 +121,7 @@ CoaxialCable.inner.live          = true
 
 CoaxialCable.dielectric.on       = true
 CoaxialCable.dielectric.radius   = 2.0
-CoaxialCable.dielectric.eps      = 5
-
+CoaxialCable.dielectric.eps      = 10
 
 # ====================================================================================================
 # FIELD_STEPPER CLASS OPTIONS
