+++ Source/AmrMesh/CD_AmrMesh.cpp
@@ -1819,6 +1819,16 @@ AmrMesh::interpToCentroids(EBAMRCellData& a_data, const std::string a_realm, con
 }
 
 void
+AmrMesh::setCoarsestGrid(const IntVect& a_nCells) {
+  CH_TIME("AmrMesh::setCoarsestGrid()");
+  if (m_verbosity > 3) {
+    pout() << "AmrMesh::setCoarsestGrid()" << endl;
+  }
+
+  m_numCells = a_nCells;
+}
+
+void
 AmrMesh::parseVerbosity()
 {
   CH_TIME("AmrMesh::parseVerbosity");
